requirements.txt
cv2
numpy
request